Remove Deprecated AdvancedClient and AdvancedServer Headers
- Deleted `AdvancedClient.h` and `AdvancedServer.h` files to streamline the codebase and eliminate unused components.
- This cleanup aligns with recent refactoring efforts to modernize the protocol and session interfaces, enhancing overall maintainability.
